+++ weblate/trans/specialchars.py
@@ -31,11 +31,13 @@ import six
 
 # Names of hardcoded chars
 CHAR_NAMES = {
-    '→': ugettext_lazy('Insert tab character'),
-    '↵': ugettext_lazy('Insert new line'),
+    '\t': ugettext_lazy('Insert tab character'),
+    '\n': ugettext_lazy('Insert new line'),
     '…': ugettext_lazy('Insert horizontal ellipsis'),
-    '।': ugettext_lazy('Danda'),
-    '॥': ugettext_lazy('Double danda'),
+}
+DISPLAY_CHARS = {
+    '\t': '↹',
+    '\n': '↵',
 }
 
 # Quotes definition for each language, based on CLDR data
@@ -438,14 +440,23 @@ def get_quote(code, data, name):
 def format_char(char):
     """Return verbose description of a character."""
     display = char
+    if char in DISPLAY_CHARS:
+        display = DISPLAY_CHARS[char]
     if char in CHAR_NAMES:
         name = CHAR_NAMES[char]
     elif unicodedata.category(char)[0] in ('C', 'Z'):
         # Various control and space chars
-        name = unicodedata.name(char)
-        display = ''.join([
-            x[0] for x in name.replace('-TO-', ' ').replace('-', ' ').split()
-        ])
+        try:
+            name = unicodedata.name(char)
+            display = ''.join([
+                x[0] for x in name.replace('-TO-', ' ').replace('-', ' ').split()
+            ])
+            name = _('Insert {0}').format(name)
+        except ValueError:
+            # Char now known to unicode data
+            # This mostly happens for control chars < 0x20
+            name = display = char.encode('unicode_escape')
+            name = _('Insert character {0}').format(name)
     else:
         name = _('Insert character {0}').format(char)
     return name, display, char