+++ docs/changes.rst
@@ -7,6 +7,7 @@ weblate 2.18
 Released on ? 2017.
 
 * Extended contributor stats.
+* Improved configuration of special chars virtual keyboard.
 
 weblate 2.17.1
 --------------